Place API actions behind interface
